requirements.txt
# PyTorch (CPU version - for GPU, install separately with CUDA)
torch>=2.0.0
torchvision>=0.15.0

# Web framework
fastapi>=0.110.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Model dependencies
einops>=0.6.0
timm>=0.9.0
kornia>=0.7.0
huggingface_hub>=0.23.0

# BiRefNet dependencies
scipy>=1.10.0
